+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## [0.1.0] [unreleased]
 ### Added
 - Added data bags
+- Added new ChefDK installation method
+- Added new config `chef_dk` - set ChefDK version or `current`
 
 ## [0.0.2] - 2016-05-20
 ### Fixed
-- Fix bug if chefDK is not installed
+- Fix bug if ChefDK is not installed
 
 ## [0.0.1] - 2016-05-20
 - Initial release